PodPluse_Backend/requirements.txt
Django
django-decouple
djangorestframework
djangorestframework-simplejwt
django-cors-headers
validate_email
pillow
openai
pyttsx3
pydub
gTTS
transformers
torch
coqui-tts